attendance-backend/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-JWT-Extended>=4.5.0
Flask-CORS>=4.0.0
opencv-python>=4.8.0
qrcode>=7.4.0
Pillow>=10.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
numpy>=1.24.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0
blinker>=1.6.0


